backend/requirements.txt
# Django Backend Requirements
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Optional but useful
PyJWT==2.8.0

# Development tools (optional)
pytest==8.1.0
pytest-django==4.7.0
pytest-cov==4.1.0

# If you use PostgreSQL later
psycopg2-binary==2.9.9
